Move settings.tsx into the drawer
